# docs/openapi/api-gateway-openapi.json
{"openapi":"3.0.0","paths":{"/appointments":{"get":{"operationId":"AppointmentsController_findAll","parameters":[],"responses":{"200":{"description":"Listado de citas."},"401":{"description":"JWT ausente o inválido."}},"security":[{"JWT":[]}],"summary":"Listar citas según el rol del usuario autenticado","tags":["Appointments"]},"post":{"operationId":"AppointmentsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAppointmentDto"}}}},"responses":{"201":{"description":"Cita creada correctamente."},"403":{"description":"Rol sin permiso para crear cita."},"409":{"description":"Horario no disponible o empalme detectado."}},"security":[{"JWT":[]}],"summary":"Crear una nueva cita odontológica","tags":["Appointments"]}},"/appointments/availability":{"get":{"operationId":"AppointmentsController_getAvailability","parameters":[{"name":"dentistId","required":true,"in":"query","schema":{"type":"string"}},{"name":"date","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Disponibilidad calculada."},"401":{"description":"JWT ausente o inválido."}},"security":[{"JWT":[]}],"summary":"Consultar disponibilidad de un dentista por fecha","tags":["Appointments"]}},"/appointments/{id}":{"get":{"operationId":"AppointmentsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Detalle de la cita."},"404":{"description":"Cita no encontrada."}},"security":[{"JWT":[]}],"summary":"Consultar una cita por ID","tags":["Appointments"]}},"/appointments/{id}/reschedule":{"patch":{"operationId":"AppointmentsController_reschedule","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RescheduleAppointmentDto"}}}},"responses":{"200":{"description":"Cita reagendada correctamente"},"403":{"description":"No tiene permiso para reagendar esta cita."},"404":{"description":"Cita no encontrada."},"409":{"description":"El nuevo horario no está disponible o genera empalme."}},"security":[{"JWT":[]}],"summary":"Cambiar fecha y hora de una cita","tags":["Appointments"]}},"/appointments/{id}/cancel":{"patch":{"operationId":"AppointmentsController_cancel","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Cita cancelada correctamente."},"403":{"description":"No tiene permiso para cancelar esta cita."}},"security":[{"JWT":[]}],"summary":"Cancelar una cita","tags":["Appointments"]}},"/appointments/{id}/confirm":{"patch":{"operationId":"AppointmentsController_confirm","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Cita confirmada correctamente."},"403":{"description":"Rol sin permiso para confirmar cita."}},"security":[{"JWT":[]}],"summary":"Confirmar una cita como dentista o administrador","tags":["Appointments"]}},"/appointments/{id}/complete":{"patch":{"operationId":"AppointmentsController_complete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Cita completada correctamente"},"403":{"description":"No tiene permiso para completar esta cita."},"404":{"description":"Cita no encontrada."}},"security":[{"JWT":[]}],"summary":"Marcar como completada una cita","tags":["Appointments"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"Usuario registrado correctamente."},"400":{"description":"Datos inválidos o usuario existente."}},"summary":"Registrar paciente en Dentia","tags":["Auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Login correcto. Devuelve accessToken y datos del usuario."},"400":{"description":"Datos inválidos."},"401":{"description":"Credenciales inválidas."}},"summary":"Iniciar sesión y obtener JWT","tags":["Auth"]}},"/profile":{"get":{"operationId":"UsersController_getProfile","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/dentists":{"get":{"operationId":"DentistsController_findAll","parameters":[],"responses":{"200":{"description":"Listado de dentistas afiliados."},"503":{"description":"auth-service no disponible."}},"summary":"Listar dentistas afiliados disponibles","tags":["Dentists"]}},"/dentists/{domainId}":{"get":{"operationId":"DentistsController_findOne","parameters":[{"name":"domainId","required":true,"in":"path","description":"ID de dominio del dentista.","schema":{"example":"d1","type":"string"}}],"responses":{"200":{"description":"Detalle del dentista afiliado."},"404":{"description":"Dentista no encontrado."},"503":{"description":"auth-service no disponible."}},"summary":"Consultar dentista por domainId","tags":["Dentists"]}},"/prescriptions":{"post":{"operationId":"PrescriptionsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePrescriptionDto"}}}},"responses":{"201":{"description":"Receta creada correctamente."},"401":{"description":"JWT ausente o inválido."},"403":{"description":"Solo administrador o dentista pueden crear recetas."},"503":{"description":"prescriptions-service no disponible."}},"security":[{"JWT":[]}],"summary":"Crear receta asociada a una cita","tags":["Prescriptions"]}},"/prescriptions/{id}":{"get":{"operationId":"PrescriptionsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"JWT":[]}],"tags":["Prescriptions"]}},"/appointments/{appointmentId}/prescriptions":{"get":{"operationId":"PrescriptionsController_findByAppointment","parameters":[{"name":"appointmentId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"JWT":[]}],"tags":["Prescriptions"]}},"/chat/conversations":{"get":{"operationId":"ChatController_listConversations","parameters":[],"responses":{"200":{"description":"Listado de conversaciones del paciente o dentista."},"401":{"description":"JWT ausente o inválido."},"503":{"description":"chat-service no disponible."}},"security":[{"JWT":[]}],"summary":"Listar conversaciones del usuario autenticado","tags":["Chat"]},"post":{"operationId":"ChatController_createConversation","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateConversationDto"}}}},"responses":{"201":{"description":"Conversación creada correctamente."},"401":{"description":"JWT ausente o inválido."},"403":{"description":"No existe relación válida paciente-dentista o no tiene permiso."},"503":{"description":"chat-service no disponible."}},"security":[{"JWT":[]}],"summary":"Crear conversación paciente-dentista","tags":["Chat"]}},"/chat/conversations/{conversationId}/messages":{"get":{"operationId":"ChatController_listMessages","parameters":[{"name":"conversationId","required":true,"in":"path","description":"ID de la conversación.","schema":{"example":"665f1b2c7a9e8a0012ab3456","type":"string"}}],"responses":{"200":{"description":"Historial de mensajes."},"401":{"description":"JWT ausente o inválido."},"403":{"description":"No tiene acceso a esta conversación."},"404":{"description":"Conversación no encontrada."},"503":{"description":"chat-service no disponible."}},"security":[{"JWT":[]}],"summary":"Listar mensajes de una conversación","tags":["Chat"]},"post":{"operationId":"ChatController_sendMessage","parameters":[{"name":"conversationId","required":true,"in":"path","description":"ID de la conversación.","schema":{"example":"665f1b2c7a9e8a0012ab3456","type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendMessageDto"}}}},"responses":{"201":{"description":"Mensaje enviado correctamente."},"401":{"description":"JWT ausente o inválido."},"403":{"description":"No tiene permiso para enviar mensajes en esta conversación."},"404":{"description":"Conversación no encontrada."},"503":{"description":"chat-service no disponible."}},"security":[{"JWT":[]}],"summary":"Enviar mensaje en una conversación","tags":["Chat"]}},"/chat/conversations/{conversationId}/read":{"patch":{"operationId":"ChatController_markAsRead","parameters":[{"name":"conversationId","required":true,"in":"path","description":"ID de la conversación.","schema":{"example":"665f1b2c7a9e8a0012ab3456","type":"string"}}],"responses":{"200":{"description":"Conversación marcada como leída."},"401":{"description":"JWT ausente o inválido."},"403":{"description":"No tiene acceso a esta conversación."},"404":{"description":"Conversación no encontrada."},"503":{"description":"chat-service no disponible."}},"security":[{"JWT":[]}],"summary":"Marcar conversación como leída","tags":["Chat"]}},"/files":{"post":{"operationId":"FilesController_upload","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/UploadFileDto"}}}},"responses":{"201":{"description":"Archivo subido correctamente."},"400":{"description":"Archivo requerido o datos inválidos."},"401":{"description":"JWT ausente o inválido."},"403":{"description":"No tiene permiso para subir este archivo."},"503":{"description":"files-service no disponible."}},"security":[{"JWT":[]}],"summary":"Subir archivo clínico","tags":["Files"]},"get":{"operationId":"FilesController_findAll","parameters":[{"name":"type","required":false,"in":"query","description":"Filtrar archivos por tipo clínico.","schema":{"example":"radiografia"}},{"name":"appointmentId","required":false,"in":"query","description":"Filtrar archivos por cita.","schema":{"example":"appointment-id"}},{"name":"patientId","required":false,"in":"query","description":"Filtrar archivos por paciente.","schema":{"example":"p-123"}}],"responses":{"200":{"description":"Listado de archivos clínicos."},"401":{"description":"JWT ausente o inválido."},"403":{"description":"No tiene permiso para consultar estos archivos."},"503":{"description":"files-service no disponible."}},"security":[{"JWT":[]}],"summary":"Listar archivos clínicos según filtros y permisos del usuario","tags":["Files"]}},"/files/{id}":{"get":{"operationId":"FilesController_findOne","parameters":[{"name":"id","required":true,"in":"path","description":"ID del archivo clínico.","schema":{"example":"665f1b2c7a9e8a0012ab3456","type":"string"}}],"responses":{"200":{"description":"Metadatos del archivo clínico."},"401":{"description":"JWT ausente o inválido."},"403":{"description":"No tiene permiso para consultar este archivo."},"404":{"description":"Archivo no encontrado."},"503":{"description":"files-service no disponible."}},"security":[{"JWT":[]}],"summary":"Consultar metadatos de un archivo clínico","tags":["Files"]},"delete":{"operationId":"FilesController_remove","parameters":[{"name":"id","required":true,"in":"path","description":"ID del archivo clínico.","schema":{"example":"665f1b2c7a9e8a0012ab3456","type":"string"}}],"responses":{"200":{"description":"Archivo eliminado correctamente."},"401":{"description":"JWT ausente o inválido."},"403":{"description":"No tiene permiso para eliminar este archivo."},"404":{"description":"Archivo no encontrado."},"503":{"description":"files-service no disponible."}},"security":[{"JWT":[]}],"summary":"Eliminar archivo clínico","tags":["Files"]}},"/files/{id}/download":{"get":{"operationId":"FilesController_download","parameters":[{"name":"id","required":true,"in":"path","description":"ID del archivo clínico.","schema":{"example":"665f1b2c7a9e8a0012ab3456","type":"string"}}],"responses":{"200":{"description":"Stream del archivo clínico."},"401":{"description":"JWT ausente o inválido."},"403":{"description":"No tiene permiso para descargar este archivo."},"404":{"description":"Archivo no encontrado."},"503":{"description":"files-service no disponible."}},"security":[{"JWT":[]}],"summary":"Descargar archivo clínico","tags":["Files"]}}},"info":{"title":"Dentia API Gateway","description":"API Gateway REST de Dentia. Centraliza autenticación, citas, chat, archivos, recetas y comunicación con microservicios.","version":"1.0.0","contact":{}},"tags":[{"name":"Auth","description":"Registro, login y sesión"},{"name":"Dentists","description":"Consulta de dentistas afiliados"},{"name":"Appointments","description":"Gestión de citas odontológicas"},{"name":"Chat","description":"Comunicación paciente-dentista"},{"name":"Files","description":"Archivos clínicos"},{"name":"Prescriptions","description":"Recetas e indicaciones"}],"servers":[],"components":{"securitySchemes":{"JWT":{"scheme":"bearer","bearerFormat":"JWT","type":"http","description":"JWT obtenido desde /auth/login"}},"schemas":{"CreateAppointmentDto":{"type":"object","properties":{"patientId":{"type":"string","example":"p-123","description":"ID de dominio del paciente"},"dentistId":{"type":"string","example":"d1","description":"ID de dominio del dentista."},"startAt":{"type":"string","example":"2026-06-01T10:00:00.000Z","description":"Fecha y hora de inicio de la cita."},"endAt":{"type":"string","example":"2026-06-01T10:30:00.000Z","description":"Fecha y hora de finalización de la cita."},"reason":{"type":"string","example":"Consulta general","description":"Motivo de la cita."},"notes":{"type":"string","example":"Paciente con sensibilidad en encías","description":"Notas adicionales."}},"required":["patientId","dentistId","startAt","endAt"]},"RescheduleAppointmentDto":{"type":"object","properties":{"startAt":{"type":"string","example":"2026-06-01T11:00:00.000Z","description":"Nueva fecha y hora de inicio de la cita en formato ISO 8601."},"endAt":{"type":"string","example":"2026-06-01T11:30:00.000Z","description":"Nueva fecha y hora de finalización de la cita en formato ISO 8601."}},"required":["startAt","endAt"]},"RegisterDto":{"type":"object","properties":{"email":{"type":"string","example":"paciente@dentia.test","description":"Correo electrónico del usuario."},"password":{"type":"string","example":"Password123","minLength":8,"description":"Contraseña del usuario."},"fullName":{"type":"string","example":"Paciente Dentia","minLength":3,"description":"Nombre completo del usuario."}},"required":["email","password","fullName"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"paciente@dentia.test","description":"Correo electrónico registrado."},"password":{"type":"string","example":"Password123","minLength":6,"description":"Contraseña del usuario."}},"required":["email","password"]},"CreatePrescriptionDto":{"type":"object","properties":{"appointmentId":{"type":"string","example":"appointment-id","description":"ID de la cita asociada a la receta."},"patientId":{"type":"string","example":"p-123","description":"ID de dominio del paciente."},"dentistId":{"type":"string","example":"d1","description":"ID de dominio del dentista."},"diagnosis":{"type":"string","example":"Caries dental leve en molar inferior.","description":"Diagnóstico registrado por el dentista."},"indications":{"type":"string","example":"Tomar analgésico cada 8 horas por 3 días y acudir a revisión.","description":"Indicaciones clínicas para el paciente."},"notes":{"type":"string","example":"Evitar alimentos muy fríos durante 48 horas.","description":"Notas adicionales de la receta."}},"required":["appointmentId","patientId","dentistId","diagnosis","indications"]},"CreateConversationDto":{"type":"object","properties":{"patientId":{"type":"string","example":"p-123","description":"ID de dominio del paciente."},"dentistId":{"type":"string","example":"d1","description":"ID de dominio del dentista."}},"required":["patientId","dentistId"]},"SendMessageDto":{"type":"object","properties":{"body":{"type":"string","example":"Hola doctor, tengo una duda sobre mi tratamiento.","description":"Contenido del mensaje."}},"required":["body"]},"UploadFileDto":{"type":"object","properties":{"file":{"type":"string","format":"binary","description":"Archivo clínico a subir."},"patientId":{"type":"string","example":"p-123","description":"ID de dominio del paciente propietario del archivo."},"appointmentId":{"type":"string","example":"appointment-id","description":"ID de la cita relacionada, si aplica."},"type":{"type":"string","example":"radiografia","description":"Tipo de archivo clínico."}},"required":["file","patientId","type"]}}}}